favourites

+++ src/main/kotlin/com/example/books/persistence/BookRepository.kt
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface BookRepository: CrudRepository<Book, Long>  {
+interface BookRepository: CrudRepository<Book, Int>  {
 }
